Moved handlers to separate files
