+++ crates/common/src/lib.rs
@@ -5,10 +5,12 @@
 //!
 //! # Modules
 //!
+//! - [`advertiser`]: Ad serving and advertiser integration functionality
 //! - [`constants`]: Application-wide constants and configuration values
 //! - [`cookies`]: Cookie parsing and generation utilities
 //! - [`error`]: Error types and error handling utilities
 //! - [`gdpr`]: GDPR consent management and TCF string parsing
+//! - [`geo`]: Geographic location utilities and DMA code extraction
 //! - [`models`]: Data models for ad serving and callbacks
 //! - [`prebid`]: Prebid integration and real-time bidding support
 //! - [`privacy`]: Privacy utilities and helpers
@@ -18,13 +20,16 @@
 //! - [`test_support`]: Testing utilities and mocks
 //! - [`why`]: Debugging and introspection utilities
 
+pub mod advertiser;
 pub mod constants;
 pub mod cookies;
 pub mod error;
 pub mod gdpr;
+pub mod geo;
 pub mod models;
 pub mod prebid;
 pub mod privacy;
+pub mod publisher;
 pub mod settings;
 pub mod synthetic;
 pub mod templates;
